requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
auto-gptq==0.2.2
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
customtkinter==5.2.0
darkdetect==0.7.1
dataclasses-json==0.5.8
datasets==2.13.1
diffusers==0.14.0
dill==0.3.6
distlib==0.3.6
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
langchain==0.0.208
langchainplus-sdk==0.0.16
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.23.3
openapi-schema-pydantic==1.2.4
packaging==21.3
pandas==2.0.2
Pillow==9.2.0
pip==23.1.2
platformdirs==3.8.0
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rouge==1.0.1
safetensors==0.3.1
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.16
sympy==1.12
tenacity==8.2.2
tk==0.1.0
tokenizers==0.12.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.64.1
transformers==4.31.0.dev0
typing_extensions==4.3.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.23.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.8.1
